move queries up to loaders
